Updating the Shopping Cart Home Page
In this latest update, I've introduced a new feature: the ability to remove items from the cart. Additionally, I've made some refinements to the output of this class to seamlessly integrate the item deletion functionality into the Shopping List. These changes not only enhance the functionality but also improve the overall aesthetics of the output.